Move AssimpCustom.h to the public folder
It's needed if the RuntimeMeshExporter.h file is used from outside of
the RuntimeMeshImportExport module.

That's because the RuntimeMeshExporter.h file includes the
AssimpCustom.h file. This means as soon as you include
RuntimeMeshExpoter.h outside of the module, the compiler will try to
find the included AssimpCustom.h file. This will fail if AssimpCustom.h
is not exposed in the public folder.

This issue probably wasn't noticed by the plugin creator because he only
used the RuntimeMeshExporter from Blueprint, which doesn't show the
issue.