Specify materials names in FExportableMeshSection
Before, the code was using the memory address of the materials to know
if a mesh was reusing a material from the scene or not. It was also
using the 'GetName()' of the UMaterialInterface to set the
AI_MATKEY_NAME attribute of the Assimp material.

Now, it is instead relying on a name that is passed in the
'materialToExport' struct within 'FExportableMeshSection'.
This change allows the code using the exporter to define the names that
the materials should receive in the 3D file.