Crash if assimp dll failed to load
Before, if the assimp dll failed to load, it would only crash when
trying to use it. This means that the problem is noticed very late and
there is no clear error message to the user.

Forcing a crash if the assimp dll failed to be loaded makes sure the
user has an error message to report to us. Detecting a problem early
makes it easier to detect and debug.